remove input crtp
this is no longer needed with the current design
