+++ src/shogun/mathematics/graph/CMakeLists.txt
@@ -12,7 +12,6 @@ set(SHOGUN_CORE_GRAPH_HEADERS
 	${CMAKE_CURRENT_SOURCE_DIR}/nodes/Input.h
 	${CMAKE_CURRENT_SOURCE_DIR}/ops/Operator.h
 	${CMAKE_CURRENT_SOURCE_DIR}/ops/abstract/BinaryOperator.h
-	${CMAKE_CURRENT_SOURCE_DIR}/ops/abstract/InputImpl.h
 	)
 
 set(SHOGUN_CORE_GRAPH_SOURCES
